Name the argument to remove NA values from the vector
The only time the expression inside any returns an NA is if the baseline or stdthr
calculations are messed up, which means this ROI has either all negative data, or
some other issues going on and should be avoided anyway. Thus, if the expression is
ever NA, na.rm makes any() false, and causes the early exit.